update to refactor the functions
refactored the function to have a separate method to display a particular frame.

+++ Source/Test/HelloWorldAddon/Marbles_Frame.py
@@ -19,6 +19,7 @@ import Wintergatan.Physics.MarblePhysics_OOP as MP
 # customise path
 #sys.path.append('C:/Users/marga/Wintergatan_git/simulation_team_physics_simulation/Source/Test/HelloWorldAddon')
 #sys.path.append('C:/vsatish/MMX/Blender/Project/GitHub/simulation_team_physics_simulation/Source/Test/HelloWorldAddon')
+#sys.path.append('C:/Program Files/Blender Foundation/Blender 2.93/2.93/scripts/addons')
 
 import Wintergatan.Test.HelloWorldAddon.helloworld as hw
 
@@ -32,43 +33,59 @@ class DisplayFrame(bpy.types.Operator):
     bl_label = "Display Frame"         # Display name in the interface.
     bl_options = {'REGISTER', 'UNDO'}  # Enable undo for the operator.
 
-    def execute(self, context):        # execute() is called when running the operator.
+    selected_frame = 20
+    x_marbles = 4
+    y_marbles = 4
+    marble_depth = 1
+    nMarbles = x_marbles*y_marbles*marble_depth
+
+    physics_output = []
+    x_coordinates = []
+    y_coordinates = []
+    z_coordinates = []
+    marble_radius = 0
+    nsteps = 1
+
+    # execute() is called when running the operator.
+    def execute(self, context): 
         
         # Deselcting all objects
         bpy.ops.object.select_all(action='DESELECT')
         
-        selected_frame = 20
-        x_marbles = 4
-        y_marbles = 4
-        marble_depth = 1
-        n_mables = x_marbles*y_marbles*marble_depth
-        
-        physics_output = MP.run_test_problem(0.4, [x_marbles, y_marbles, marble_depth])
-        x_coordinates, y_coordinates, z_coordinates, marble_radius, nsteps = physics_output
-        
-
-        
-        scene = context.scene
-        objs = scene.objects
-        for obj in objs:
-
-            bpy.data.objects
-            bpy.data.objects[obj.name].select_set(True)
-            bpy.ops.object.delete()
-                        
-        for i in range(n_mables):
-            print('n = ' + str(i))
-            print('x = ' + str(x_coordinates[i][selected_frame]))
-            print('y = ' + str(y_coordinates[i][selected_frame]))
-            print('z = ' + str(z_coordinates[i][selected_frame]))
-            
-            try:
-                bpy.ops.mesh.primitive_uv_sphere_add(segments=32, ring_count=16, radius=marble_radius, calc_uvs=True, enter_editmode=False, align='WORLD', location=(x_coordinates[i][selected_frame], y_coordinates[i][selected_frame], z_coordinates[i][selected_frame]), rotation=(0.0, 0.0, 0.0), scale=(1.0, 1.0, 1.0))
-                #bpy.ops.mesh.primitive_cube_add(enter_editmode=False, align='WORLD', location=(x_coordinates[i][selected_fame], y_coordinates[i][selected_fame], z_coordinates[i][selected_fame]), scale=(1, 1, 1))
-            except:
-                print('marble could not be created')
+        dispFrame(self, context)
 
         return {'FINISHED'}            # Lets Blender know the operator finished successfully.
+
+def dispFrame(self, context):
+    scene = context.scene
+    objs = scene.objects
+
+    print(DisplayFrame.selected_frame)
+    print(DisplayFrame.physics_output)
+    print(DisplayFrame.nMarbles)
+    x_coordinates, y_coordinates, z_coordinates, marble_radius, nsteps = DisplayFrame.physics_output
+
+    if (nsteps < 1):
+        return
+    
+    for obj in objs:
+
+        bpy.data.objects
+        bpy.data.objects[obj.name].select_set(True)
+        bpy.ops.object.delete()
+    
+    for i in range(DisplayFrame.nMarbles):
+        print('n = ' + str(i))
+        print('x = ' + str(x_coordinates[i][DisplayFrame.selected_frame]))
+        print('y = ' + str(y_coordinates[i][DisplayFrame.selected_frame]))
+        print('z = ' + str(z_coordinates[i][DisplayFrame.selected_frame]))
+        
+        try:
+            bpy.ops.mesh.primitive_uv_sphere_add(segments=32, ring_count=16, radius=marble_radius, calc_uvs=True, enter_editmode=False, align='WORLD', location=(x_coordinates[i][DisplayFrame.selected_frame], y_coordinates[i][DisplayFrame.selected_frame], z_coordinates[i][DisplayFrame.selected_frame]), rotation=(0.0, 0.0, 0.0), scale=(1.0, 1.0, 1.0))
+            #bpy.ops.mesh.primitive_cube_add(enter_editmode=False, align='WORLD', location=(x_coordinates[i][selected_fame], y_coordinates[i][selected_fame], z_coordinates[i][selected_fame]), scale=(1, 1, 1))
+        except:
+            print('marble could not be created')
+    return
 
 def init_setup():
     # initial settings for operation
@@ -91,6 +108,11 @@ def menu_func(self, context):
 
 def register():
     init_setup()
+
+    # moved physics sim code here to startup to isolate for UI and flow
+    DisplayFrame.physics_output = MP.run_test_problem(0.4, [DisplayFrame.x_marbles, DisplayFrame.y_marbles, DisplayFrame.marble_depth])
+    DisplayFrame.selected_frame = 20 #(nsteps + 1) / 2 # display the middle frame
+
     bpy.utils.register_class(DisplayFrame)
     bpy.types.VIEW3D_MT_object.append(menu_func)  # Adds the new operator to an existing menu.
 
@@ -98,6 +120,17 @@ def unregister():
     bpy.types.VIEW3D_MT_object.remove(menu_func)
     bpy.utils.unregister_class(DisplayFrame)
 
+def __init__(self):
+    DisplayFrame.x_marbles = 4
+    DisplayFrame.y_marbles = 4
+    DisplayFrame.marble_depth = 1
+    DisplayFrame.nMarbles = DisplayFrame.x_marbles*DisplayFrame.y_marbles*DisplayFrame.marble_depth
+    #self.physics_output = []
+    self.x_coordinates = []
+    self.y_coordinates = []
+    self.z_coordinates = []
+    self.marble_radius = 0
+    self.nsteps = 1
 
 # This allows you to run the script directly from Blender's Text editor
 # to test the add-on without having to install it.
